Moved execution and output methods out of main
